Resolve an issue regarding repeated filenames
Considering that the archive is named based on the date it was created, a conflict will occur when tasked to archive multiple threads that were made on the same date.
To resolve this, the code emulates how OS resolves multiple pastes of the same files in the directory.

Note: This also allows duplicates in terms of content. Since 'publish_to_csv' returns a filename, the caller 'Archive' must immediately reference it.